+++ src/Minefields/Utils.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 #include <cstdlib>
-#include <uniTestingcpp/Utils.h>
+#include <Minefields/Utils.h>
 #include <string>
